Fix data load syncrhonization bug where a component can load data before authentication state is finalized
