+++ services/agora/src/stores/authentication.ts
@@ -5,10 +5,12 @@ export const useAuthenticationStore = defineStore("authentication", () => {
   const verificationPhoneNumber = ref("");
   const verificationDefaultCallingCode = ref("");
   const isAuthenticated = ref(false);
+  const isAuthInitialized = ref(false);
 
   return {
     isAuthenticated,
     verificationPhoneNumber,
     verificationDefaultCallingCode,
+    isAuthInitialized,
   };
 });
